이슈 Junit 테스트 분리 + issueCreation 테스트 수정
